Added CMakeLists.txt and more structure to source files
