+++ .gitignore
@@ -343,3 +343,6 @@ healthchecksdb
 
 # Backup folder for Package Reference Convert tool in Visual Studio 2017
 MigrationBackup/
+
+# Specific to this report
+results/*
