+++ src/Sample_FundamentalEstimator/line_estimator.h
@@ -1,5 +1,3 @@
-#include "stdafx.h"
-
 #include <iostream>
 #include <math.h>
 #include <chrono>
